Add files via upload
This DB was created in order to give the Human Resources department a single place where they could record and track every training session the company has given to the employees. It provides different reports for the department according to their needs. The DB tracks all kinds of training sessions data, calculates and processes it for later use in reports. All outputs are formatted to fit ISO 9001 requirements, since the company certificates said norm.